feat: event unpack
